+++ source/1_engine/rex_engine/include/rex_engine/filesystem/path.h
@@ -6,5 +6,5 @@ namespace rex
 }
 
 #ifdef REX_PLATFORM_WINDOWS
-#include "rex_engine/platform/win/filesystem/path.h"
+#include "rex_engine/platform/win/filesystem/win_path.h"
 #endif
