+++ source/1_engine/rex_engine/src/filesystem/path.cpp
@@ -1,0 +1,6 @@
+#include "rex_engine/filesystem/path.h"
+
+namespace rex
+{
+    
+}
